+++ lib/bootstrap_form_extensions.rb
@@ -5,6 +5,7 @@ require 'bootstrap_form_extensions/arrayed_field'
 require 'bootstrap_form_extensions/scheduler'
 require 'bootstrap_form_extensions/date_time_pickers'
 require 'bootstrap-datepicker-rails'
+require 'bootstrap_form_extensions/select_or_new'
 
 module BootstrapFormExtensions
   module Rails
@@ -17,4 +18,5 @@ BootstrapForm::FormBuilder.send :include, BootstrapFormExtensions::SubmitBar,
                                           BootstrapFormExtensions::Timespan,
                                           BootstrapFormExtensions::ArrayedField,
                                           BootstrapFormExtensions::Scheduler,
-                                          BootstrapFormExtensions::DateTimePickers
+                                          BootstrapFormExtensions::DateTimePickers,
+                                          BootstrapFormExtensions::SelectOrNew