+++ test/dummy/app/models/category.rb
@@ -1,0 +1,7 @@
+class Category < ActiveRecord::Base
+
+  has_many :things
+
+  validates :name, presence: true
+
+end