+++ test/dummy/db/migrate/20151106165522_add_category.rb
@@ -1,0 +1,10 @@
+class AddCategory < ActiveRecord::Migration
+
+  def change
+    create_table :categories do |t|
+      t.string :name
+    end
+    add_column :things, :category_id, :integer, references: 'categories'
+  end
+
+end