+++ CHANGELOG.md
@@ -4,6 +4,7 @@ Features:
 
   - Add SubmitBar control
   - Add Timespan control
-  - Add ArrayedFields
+  - Add ArrayedField
   - Add Scheduler
   - Add Date and Time Pickers
+  - Add Select or new control